Made jump work properly, and added custom gravity scale, changed player color to green,
im still feeling sick, but it's just been a runny nose right now. after this, we should focus on making a little area with a wide variety of platforms for the player to jump around on and have the camera follow the player. that way, we can think of how best to approach the game feel of this (like how moving and jumping feels) and make changes to that. because if we try to change that stuff later, it will be a lot harder and we'll end up having to rework levels entirely.