added cinemachine + camera controller for player
i added a unity package called cinemachine, which makes it easier to make a custom camera to follow the player. i also added a bunch of various platforms for us to test out how the player jumps around. there's a few bugs and features i already noticed that should be fixed/added, so we can work on some of those, and then after that add some more of the player controls that need to be added and some basic enemy AI