added player physics material to prevent collision bug & freezed rotation for player
i added an isJumping boolean when the player gets stuck to a ceiling, but i also found out that the bug happens to the sides of platforms too. I learned that an easy solution for this is to give the player a physics 2d material and set their friction to 0, so they just slide seamlessly now! also, i froze the rotation of the player across the Z axis since they shouldn't need to use that for a 2d platformer anyways. still need to add coyote time and some other stuff to enhance the jump physics, but this is looking really good!